Fix: Use jQuery promise methods instead of native Promise
The bug was that createAjaxRequest was using .then() which returns a native Promise, but createFetchHandler was calling .done() and .fail() which only exist on jQuery Promises. Changed to use .done()/.fail()/.always() to maintain jQuery Promise compatibility.

This fixes the issue where API responses were successful but Redux state wasn't being updated, causing the app to remain stuck on the loading screen.